+++ examples/libp2p-in-the-browser/webtransport/fetch-file-from-kubo/vite.config.js
@@ -1,8 +1,0 @@
-export default {
-  build: {
-    target: 'es2020'
-  },
-  optimizeDeps: {
-    esbuildOptions: { target: 'es2020', supported: { bigint: true } }
-  }
-}