+++ packages/kad-dht/src/query-self.ts
@@ -1,4 +1,4 @@
-import { setMaxListeners } from 'events'
+import { setMaxListeners } from '@libp2p/interface/events'
 import { logger, type Logger } from '@libp2p/logger'
 import { anySignal } from 'any-signal'
 import length from 'it-length'
@@ -105,24 +105,24 @@ export class QuerySelf implements Startable {
 
     this.querySelfPromise = pDefer()
 
-    if (this.routingTable.size === 0) {
-      // wait to discover at least one DHT peer
-      await pEvent(this.routingTable, 'peer:add')
-    }
-
     if (this.started) {
       this.controller = new AbortController()
       const signal = anySignal([this.controller.signal, AbortSignal.timeout(this.queryTimeout)])
 
       // this controller will get used for lots of dial attempts so make sure we don't cause warnings to be logged
-      try {
-        if (setMaxListeners != null) {
-          setMaxListeners(Infinity, signal)
-        }
-      } catch {} // fails on node < 15.4
+      setMaxListeners(Infinity, signal)
 
       try {
+        if (this.routingTable.size === 0) {
+          this.log('routing table was empty, waiting for some peers before running query')
+          // wait to discover at least one DHT peer
+          await pEvent(this.routingTable, 'peer:add', {
+            signal
+          })
+        }
+
         this.log('run self-query, look for %d peers timing out after %dms', this.count, this.queryTimeout)
+        const start = Date.now()
 
         const found = await pipe(
           this.peerRouting.getClosestPeers(this.components.peerId.toBytes(), {
@@ -133,16 +133,16 @@ export class QuerySelf implements Startable {
           async (source) => length(source)
         )
 
-        this.log('self-query ran successfully - found %d peers', found)
+        this.log('self-query found %d peers in %dms', found, Date.now() - start)
+      } catch (err: any) {
+        this.log.error('self-query error', err)
+      } finally {
+        signal.clear()
 
         if (this.initialQuerySelfHasRun != null) {
           this.initialQuerySelfHasRun.resolve()
           this.initialQuerySelfHasRun = undefined
         }
-      } catch (err: any) {
-        this.log.error('self-query error', err)
-      } finally {
-        signal.clear()
       }
     }