+++ packages/libp2p/src/circuit-relay/server/index.ts
@@ -1,5 +1,4 @@
-import { setMaxListeners } from 'events'
-import { EventEmitter } from '@libp2p/interface/events'
+import { TypedEventEmitter, setMaxListeners } from '@libp2p/interface/events'
 import { logger } from '@libp2p/logger'
 import { peerIdFromBytes } from '@libp2p/peer-id'
 import { RecordEnvelope } from '@libp2p/peer-record'
@@ -99,7 +98,7 @@ const defaults = {
   maxOutboundStopStreams: MAX_CONNECTIONS
 }
 
-class CircuitRelayServer extends EventEmitter<RelayServerEvents> implements Startable, CircuitRelayService {
+class CircuitRelayServer extends TypedEventEmitter<RelayServerEvents> implements Startable, CircuitRelayService {
   private readonly registrar: Registrar
   private readonly peerStore: PeerStore
   private readonly addressManager: AddressManager
@@ -134,10 +133,7 @@ class CircuitRelayServer extends EventEmitter<RelayServerEvents> implements Star
     this.maxOutboundHopStreams = init.maxOutboundHopStreams
     this.maxOutboundStopStreams = init.maxOutboundStopStreams ?? defaults.maxOutboundStopStreams
 
-    try {
-      // fails on node < 15.4
-      setMaxListeners?.(Infinity, this.shutdownController.signal)
-    } catch { }
+    setMaxListeners(Infinity, this.shutdownController.signal)
 
     if (init.advertise != null && init.advertise !== false) {
       this.advertService = new AdvertService(components, init.advertise === true ? undefined : init.advertise)