+++ packages/libp2p/test/connection-manager/direct.node.ts
@@ -6,7 +6,7 @@ import path from 'node:path'
 import { yamux } from '@chainsafe/libp2p-yamux'
 import { type Connection, type ConnectionProtector, isConnection } from '@libp2p/interface/connection'
 import { AbortError } from '@libp2p/interface/errors'
-import { EventEmitter } from '@libp2p/interface/events'
+import { TypedEventEmitter } from '@libp2p/interface/events'
 import { start, stop } from '@libp2p/interface/startable'
 import { mockConnection, mockConnectionGater, mockDuplex, mockMultiaddrConnection, mockUpgrader } from '@libp2p/interface-compliance-tests/mocks'
 import { mplex } from '@libp2p/mplex'
@@ -22,7 +22,7 @@ import { pipe } from 'it-pipe'
 import { pushable } from 'it-pushable'
 import pDefer from 'p-defer'
 import pWaitFor from 'p-wait-for'
-import sinon from 'sinon'
+import Sinon from 'sinon'
 import { stubInterface } from 'sinon-ts'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import { DefaultAddressManager } from '../../src/address-manager/index.js'
@@ -49,16 +49,16 @@ describe('dialing (direct, TCP)', () => {
   let remoteAddr: Multiaddr
   let remoteComponents: Components
   let localComponents: Components
-  let resolver: sinon.SinonStub<[Multiaddr], Promise<string[]>>
+  let resolver: Sinon.SinonStub<[Multiaddr], Promise<string[]>>
 
   beforeEach(async () => {
-    resolver = sinon.stub<[Multiaddr], Promise<string[]>>()
+    resolver = Sinon.stub<[Multiaddr], Promise<string[]>>()
     const [localPeerId, remotePeerId] = await Promise.all([
       createEd25519PeerId(),
       createEd25519PeerId()
     ])
 
-    const remoteEvents = new EventEmitter()
+    const remoteEvents = new TypedEventEmitter()
     remoteComponents = defaultComponents({
       peerId: remotePeerId,
       events: remoteEvents,
@@ -66,7 +66,7 @@ describe('dialing (direct, TCP)', () => {
       upgrader: mockUpgrader({ events: remoteEvents }),
       connectionGater: mockConnectionGater(),
       transportManager: stubInterface<TransportManager>({
-        getAddrs: []
+        getAddrs: Sinon.stub().returns([])
       })
     })
     remoteComponents.peerStore = new PersistentPeerStore(remoteComponents)
@@ -78,7 +78,7 @@ describe('dialing (direct, TCP)', () => {
     remoteTM = remoteComponents.transportManager = new DefaultTransportManager(remoteComponents)
     remoteTM.add(tcp()())
 
-    const localEvents = new EventEmitter()
+    const localEvents = new TypedEventEmitter()
     localComponents = defaultComponents({
       peerId: localPeerId,
       events: localEvents,
@@ -109,7 +109,7 @@ describe('dialing (direct, TCP)', () => {
   })
 
   afterEach(() => {
-    sinon.restore()
+    Sinon.restore()
   })
 
   it('should be able to connect to a remote node via its multiaddr', async () => {
@@ -194,7 +194,7 @@ describe('dialing (direct, TCP)', () => {
 
     const dialer = new DialQueue(localComponents)
 
-    sinon.spy(localTM, 'dial')
+    Sinon.spy(localTM, 'dial')
     const connection = await dialer.dial(peerId)
     expect(localTM.dial).to.have.property('callCount', remoteAddrs.length)
     expect(connection).to.exist()
@@ -207,7 +207,7 @@ describe('dialing (direct, TCP)', () => {
       dialTimeout: 50
     })
 
-    sinon.stub(localTM, 'dial').callsFake(async (addr, options = {}) => {
+    Sinon.stub(localTM, 'dial').callsFake(async (addr, options = {}) => {
       expect(options.signal).to.exist()
       expect(options.signal?.aborted).to.equal(false)
       expect(addr.toString()).to.eql(remoteAddr.toString())
@@ -235,7 +235,7 @@ describe('dialing (direct, TCP)', () => {
     })
 
     const deferredDial = pDefer<Connection>()
-    const transportManagerDialStub = sinon.stub(localTM, 'dial')
+    const transportManagerDialStub = Sinon.stub(localTM, 'dial')
     transportManagerDialStub.callsFake(async () => deferredDial.promise)
 
     // Perform 3 multiaddr dials
@@ -273,7 +273,7 @@ describe('dialing (direct, TCP)', () => {
       maxParallelDialsPerPeer: 10
     })
 
-    const transportManagerDialStub = sinon.stub(localTM, 'dial')
+    const transportManagerDialStub = Sinon.stub(localTM, 'dial')
     transportManagerDialStub.callsFake(async (ma) => {
       await delay(10)
       return mockConnection(mockMultiaddrConnection(mockDuplex(), remoteComponents.peerId))
@@ -340,7 +340,7 @@ describe('libp2p.dialer (direct, TCP)', () => {
   })
 
   afterEach(async () => {
-    sinon.restore()
+    Sinon.restore()
 
     if (libp2p != null) {
       await libp2p.stop()
@@ -515,7 +515,7 @@ describe('libp2p.dialer (direct, TCP)', () => {
       connectionProtector: () => protector
     })
 
-    const protectorProtectSpy = sinon.spy(protector, 'protect')
+    const protectorProtectSpy = Sinon.spy(protector, 'protect')
 
     remoteLibp2p.components.connectionProtector = preSharedKey({ psk: swarmKeyBuffer })()
 
@@ -589,7 +589,7 @@ describe('libp2p.dialer (direct, TCP)', () => {
 
     const dials = 10
     const error = new Error('Boom')
-    sinon.stub(libp2p.components.transportManager, 'dial').callsFake(async () => Promise.reject(error))
+    Sinon.stub(libp2p.components.transportManager, 'dial').callsFake(async () => Promise.reject(error))
 
     await libp2p.peerStore.patch(remotePeerId, {
       multiaddrs: remoteLibp2p.getMultiaddrs()